refactor: replace image path
